lab3.1

+++ Lab1_DataTypes/Lab1_DataTypes.cpp
@@ -73,5 +73,5 @@ int main(){
     int m, l;
     std::cout << "Enter the coordinates of a line segment: ";   
     std::cin >> m >> l;
-    std::cout << "The midpoint of the line is at the point: " << double(m + l) / 2 << "\n";
+    std::cout << "The midpoint of the line is at the point: " << (double(m) + double(l)) / 2 << "\n";
 }
